+++ modkit/types/sandkit/api/collector.d.ts
@@ -1,0 +1,9 @@
+import { CellCoordinates } from "../../shared/player";
+
+export namespace collector {
+  export function getValueFromCellId(cellId: number): number;
+  export function getValueByType(elementType: number): number;
+  export function isCellIdCollectable(cellId: number): boolean;
+  export function isCellIdCollectableForSprite(cellId: number): boolean;
+  export function notifyPickupAtCell(...args: CellCoordinates): void;
+}